fdtdump: Discourage use of fdtdump
fdtdump is, and always has been, a quick-and-dirty debugging tool.  However
I keep getting reports of people using it for real work.  For production
decompiling of a dtb, dtc in -I dtb -O dts mode is the right tool. In the
hopes of getting that message out there, add a warning message to fdtdump
to discourage its use.

Signed-off-by: David Gibson <david@gibson.dropbear.id.au>